feat(#934): remove dependencies package
