--- eo-maven-plugin/src/main/java/org/eolang/maven/dependencies/Dependencies.java
@@ -1,9 +1,0 @@
-package org.eolang.maven.dependencies;
-
-import java.util.List;
-import org.apache.maven.model.Dependency;
-
-public interface Dependencies {
-
-    List<Dependency> toList();
-}
